Initial implementation of CCA scheme. Tested with msc98-ip.
Will be changed to have a much lighter node attachment, which has a reference to parent node attachments. 
Cumulative list information for each leaf needs only be calculated for surviving leafs when it comes time to farm. Also maintain an index of leafs which is updated at farming time. This allows us to only calculate cumulative values for previously unknown leafs.